Rename phpunit.xml -> phpunit.xml.dist.
The phpunit.xml.dist file is intended to be committed to version control and distributed with packages. This allows developers to create a (usually untracked) `phpunit.xml` file with necessary local overrides.

See [the phpunit docs](https://phpunit.readthedocs.io/en/9.0/textui.html), specifically the section for `--configuration` for reference.